dfchronicles/requirements.txt
Django==4.2.4
requests==2.26.0
psycopg==3.1.8
django-cors-headers==4.2.0
djangorestframework==3.14.0